requirements.txt
flask
flask-socketio
eventlet
python-vlc